First attempt
Given an array A[ ] of size N of  integers, find the index of values that satisfy A + B = C + D where A,B,C & D are integers values in the array.

Input
The first line of input consists number of the test cases T.
The first line of each test case should contain N, denoting the size of the array. The second line of each test case should contain (N-1) elements of the array A[ ].

Output
Print out the four numbers in the order : A B C D 

If no such elements are present ,then print : no pairs 

If there are more than one solutions, 

then return the tuple of values which are lexicographical smallest.